+++ package.json
@@ -26,6 +26,7 @@
     "@sendgrid/mail": "^8.1.3",
     "@types/bootstrap": "^5.2.10",
     "bootstrap": "^5.3.3",
+    "bootstrap-icons": "^1.11.3",
     "express": "^4.18.2",
     "resend": "^3.3.0",
     "rxjs": "~7.8.0",
