+++ src/app/component/header/header.component.ts
@@ -12,28 +12,5 @@ import { FooterComponent } from '../footer/footer.component';
 export class HeaderComponent {
   title = 'Tienda Virtual';
 
-  constructor(private resendEmailService: ResendEmailService) {}
-
-  sendEmail(event: Event) {
-    event.preventDefault();
-    const form = event.target as HTMLFormElement;
-
-    const to = (form.querySelector('#recipientEmail') as HTMLInputElement)
-      ?.value;
-    const subject = (form.querySelector('#emailSubject') as HTMLInputElement)
-      ?.value;
-    const message = (form.querySelector('#emailMessage') as HTMLTextAreaElement)
-      ?.value;
-
-    this.resendEmailService.sendEmail(to, subject, message).then(
-      () => {
-        console.log('Email sent successfully!');
-        // Aquí puedes agregar una notificación o mensaje de éxito si lo deseas
-      },
-      (error) => {
-        console.error('Failed to send email:', error);
-        // Aquí puedes manejar el error, mostrar mensajes de error, etc.
-      }
-    );
-  }
+  
 }
